+++ packages/e2e-rspack/tests/utils/types.ts
@@ -7,6 +7,7 @@ export interface E2EPlaywrightContext {
   browser?: Browser
   page?: Page
   rspackServer?: ResultPromise
+  rspackServerUrl?: string
 }
 
 export type EvaluateType = 'style' | 'attribute' | 'textContent'
@@ -14,4 +15,5 @@ export type EvaluateType = 'style' | 'attribute' | 'textContent'
 export interface Evaluator {
   getTextContent: (selector: string, page?: Page) => Promise<string | Nil>
   getColor: (selector: string, page?: Page) => Promise<string | Nil>
+  getAttribute: (selector: string, attribute: string, page?: Page) => Promise<string | Nil>
 }