+++ packages/language-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @vue-vine/language-server
 
+## 1.7.11
+
+### Patch Changes
+
+- Refactor and ready to release (beta) Rspack support.
+- Updated dependencies
+  - @vue-vine/compiler@1.7.11
+  - @vue-vine/language-service@1.7.11
+
 ## 1.7.10
 
 ### Patch Changes